+++ src/app/admin/components/admin-home/admin-combos/admin-combos.component.ts
@@ -65,7 +65,13 @@ export class AdminCombosComponent {
     const comboUser = this.comboForm.get('comboUser')?.value;
     const totalAmount = this.comboForm.get('totalAmount')?.value;
 
-    this.comboService.addCombo({ comboName, comboUser, totalAmount } as Combo)
+    const newCombo: Combo = {
+      comboName: comboName,
+      comboUser: comboUser,
+      totalAmount: totalAmount
+    };
+
+    this.comboService.addCombo(newCombo)
       .subscribe(combo => {
         this.combos.push(combo);
         this.getCombos();
@@ -74,8 +80,15 @@ export class AdminCombosComponent {
   }
 
   deleteCombo(combo: Combo): void {
-    this.combos = this.combos.filter(i => i !== combo);
-    this.comboService.deleteCombo(combo.comboId).subscribe();
+
+    if (combo.comboId) {
+      this.combos = this.combos.filter(i => i !== combo);
+      this.comboService.deleteCombo(combo.comboId).subscribe();
+    } else {
+      console.log("Không có comboId");
+    }
+
+
   }
 
   openComboDetailModal(combo: Combo) {
