+++ src/app/admin/components/admin-home/admin-inventory-branches/admin-inventory-branches.component.ts
@@ -105,7 +105,14 @@ export class AdminInventoryBranchesComponent {
     const ingredientId = this.inventoryBranchForm.get('ingredientId')?.value;
     const supplierId = this.inventoryBranchForm.get('supplierId')?.value;
     const restaurantId = this.inventoryBranchForm.get('restaurantId')?.value;
-    this.inventoryBranchService.addInventoryBranch({ ingredientId, supplierId, restaurantId } as InventoryBranch)
+
+    const newInventoryBranch: InventoryBranch = {
+      ingredientId: ingredientId,
+      supplierId: supplierId,
+      restaurantId: restaurantId,
+    };
+
+    this.inventoryBranchService.addInventoryBranch(newInventoryBranch)
       .subscribe(inventoryBranch => {
         this.inventoryBranches.push(inventoryBranch);
         this.inventoryBranchForm.reset();
@@ -113,8 +120,15 @@ export class AdminInventoryBranchesComponent {
   }
 
   deleteInventoryBranch(inventoryBranch: InventoryBranch): void {
-    this.inventoryBranches = this.inventoryBranches.filter(c => c !== inventoryBranch);
-    this.inventoryBranchService.deleteInventoryBranch(inventoryBranch.inventoryBranchId).subscribe();
+
+    if (inventoryBranch.inventoryBranchId) {
+      this.inventoryBranches = this.inventoryBranches.filter(c => c !== inventoryBranch);
+      this.inventoryBranchService.deleteInventoryBranch(inventoryBranch.inventoryBranchId).subscribe();
+    } else {
+      console.log("Không có inventoryBranchId");
+    }
+
+
   }
 
   openInventoryBranchDetailModal(inventoryBranch: InventoryBranch) {
