+++ src/app/admin/components/admin-home/admin-table-statuses/admin-table-statuses.component.ts
@@ -62,7 +62,12 @@ export class AdminTableStatusesComponent {
 
   addTableStatus(): void {
     const name = this.tableStatusForm.get('name')?.value?.trim() ?? '';
-    this.tableStatusService.addTableStatus({ name } as TableStatus)
+
+    const newTableStatus: TableStatus = {
+      name: name,
+    }
+
+    this.tableStatusService.addTableStatus(newTableStatus)
       .subscribe(tableStatus => {
         this.tableStatuses.push(tableStatus);
         this.tableStatusForm.reset();
@@ -70,8 +75,14 @@ export class AdminTableStatusesComponent {
   }
 
   deleteTableStatus(tableStatus: TableStatus): void {
-    this.tableStatuses = this.tableStatuses.filter(i => i !== tableStatus);
-    this.tableStatusService.deleteTableStatus(tableStatus.tableStatusId).subscribe();
+
+    if (tableStatus.tableStatusId) {
+      this.tableStatuses = this.tableStatuses.filter(i => i !== tableStatus);
+      this.tableStatusService.deleteTableStatus(tableStatus.tableStatusId).subscribe();
+    } else {
+      console.log("Không có tableStautsId");
+    }
+
   }
 
   openTableStatusDetailModal(tableStatus: TableStatus) {
